+++ geotrek/signage/tests/test_parsers.py
@@ -37,7 +37,39 @@ class SignageGeotrekParserTests(GeotrekParserTestMixin, TestCase):
 
     @classmethod
     def setUpTestData(cls):
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            cls.path = PathFactory.create()
         cls.filetype = FileType.objects.create(type="Photographie")
+
+    @skipIf(
+        not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only"
+    )
+    @mock.patch("requests.get")
+    @mock.patch("requests.head")
+    def test_import_cmd_raises_error_when_no_path(self, mocked_head, mocked_get):
+        self.mock_time = 0
+        self.mock_json_order = [
+            ("signage", "structure.json"),
+            ("signage", "signage_sealing.json"),
+            ("signage", "signage_conditions.json"),
+            ("signage", "signage_type.json"),
+        ]
+        # Mock GET
+        mocked_get.return_value.status_code = 200
+        mocked_get.return_value.json = self.mock_json
+        mocked_get.return_value.content = b""
+        mocked_head.return_value.status_code = 200
+
+        self.path.delete()
+        with self.assertRaisesRegex(
+            CommandError,
+            "You need to add a network of paths before importing 'Signage' objects",
+        ):
+            call_command(
+                "import",
+                "geotrek.signage.tests.test_parsers.TestGeotrekSignageParser",
+                verbosity=0,
+            )
 
     @mock.patch("requests.get")
     @mock.patch("requests.head")
@@ -71,6 +103,15 @@ class SignageGeotrekParserTests(GeotrekParserTestMixin, TestCase):
         self.assertEqual(str(signage.sealing), "Socle béton")
         conditions = [str(c.label) for c in signage.conditions.all()]
         self.assertEqual(conditions, ["Dégradé"])
+
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            signage_path = signage.topo_object.paths.get()
+            self.assertEqual(signage_path, self.path)
+            self.assertEqual(signage.topo_object.kind, "SIGNAGE")
+            self.assertAlmostEqual(signage.topo_object.offset, 430191.617, places=2)
+
+        self.assertEqual(signage.geom.geom_type, "Point")
+        self.assertEqual(signage.geom.srid, settings.SRID)
         self.assertAlmostEqual(signage.geom.x, 572941.1308660918, places=5)
         self.assertAlmostEqual(signage.geom.y, 6189000.155980503, places=5)
 
@@ -116,7 +157,8 @@ class OpenStreetMapSignageParserTests(TestCase):
     def test_import_cmd_raises_error_when_no_path(self):
         self.path.delete()
         with self.assertRaisesRegex(
-            CommandError, "You need to add a path network before importing signage"
+            CommandError,
+            "You need to add a network of paths before importing 'Signage' objects",
         ):
             call_command(
                 "import",
@@ -135,7 +177,7 @@ class OpenStreetMapSignageParserTests(TestCase):
     @skipIf(
         not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only"
     )
-    def test_topology_point(self):
+    def test_topology_is_correct(self):
         signage = self.objects.get(eid="N7872800265")
         self.assertAlmostEqual(signage.topo_object.offset, 0.0)
         self.assertEqual(signage.topo_object.paths.count(), 1)
@@ -143,7 +185,9 @@ class OpenStreetMapSignageParserTests(TestCase):
         self.assertEqual(signage_path, self.path)
         self.assertEqual(signage.topo_object.kind, "SIGNAGE")
 
-    def test_topology_point_no_dynamic_segmentation(self):
+    def test_geometry_is_correct(self):
         signage = self.objects.get(eid="N7872800265")
+        self.assertEqual(signage.geom.geom_type, "Point")
+        self.assertEqual(signage.geom.srid, settings.SRID)
         self.assertAlmostEqual(signage.geom.x, 966634.858, places=2)
         self.assertAlmostEqual(signage.geom.y, 6420758.255, places=2)